修改Badge+Slider排版 新增DocTitle.jsx and DocSubtitle.jsx
