+++ example/codes/SliderInfo.js
@@ -1,12 +1,12 @@
 
 /* Default */
-<MDL.Slider min={0}	max={100}/>
+<Slider min={0}	max={100}/>
 
 /* Starting with Initial Value && step */
-<MDL.Slider min={0}	max={100}	defaultValue={10}	step={5}/>
+<Slider min={0}	max={100}	defaultValue={10}	step={5}/>
 
 /* Disabled */
-<MDL.Slider min={0}	max={100}	defaultValue={80}	disabled={true}/>
+<Slider min={0}	max={100}	defaultValue={80}	disabled={true}/>
 
 /* Get Value */
-<MDL.Slider min={0}	max={100}	defaultValue={40}	step={10}/>
+<Slider min={0}	max={100}	defaultValue={40}	step={10}/>
